chore(certificatemanager): convert tests to Jest (#9290)
Converted all tests in the module to use Jest instead of nodeunit.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*
