+++ packages/@aws-cdk/aws-certificatemanager/.gitignore
@@ -14,5 +14,6 @@ nyc.config.js
 .LAST_PACKAGE
 *.snk
 !.eslintrc.js
+!jest.config.js
 
 junit.xml
