+++ packages/@aws-cdk/aws-certificatemanager/.npmignore
@@ -26,3 +26,4 @@ tsconfig.json
 # exclude cdk artifacts
 **/cdk.out
 junit.xml
+jest.config.js
