+++ packages/@aws-cdk/aws-certificatemanager/jest.config.js
@@ -1,0 +1,2 @@
+const baseConfig = require('../../../tools/cdk-build-tools/config/jest.config');
+module.exports = baseConfig;
